+++ tests/coretests.fs/coretests.fs.fsproj
@@ -21,6 +21,9 @@
     <Compile Include="Stocks\Services\SMAAnalysisTests.fs" />
     <Compile Include="Stocks\Services\VolumeAnalysisTests.fs" />
     <Compile Include="Stocks\Services\PositionAnalysisTests.fs" />
+    <Compile Include="Stocks\Services\Trading\TradingStrategyRunnerTests.fs" />
+    <Compile Include="Stocks\Services\Trading\ProfitPointTests.fs" />
+    <Compile Include="Stocks\TradingPerformanceViewTests.fs" />
     <Compile Include="Program.fs" />
   </ItemGroup>
